add flag to exclude kafka
