+++ beacon-chain/db/BUILD.bazel
@@ -1,11 +1,25 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
+#  Build with --define=kafka_enabled=false to exclude kafka wrapper.
+config_setting(
+    name = "kafka_disabled",
+    values = {"define": "kafka_enabled=false"},
+)
+
+# gazelle:ignore db.go db_kafka_wrapped.go
 go_library(
     name = "go_default_library",
     srcs = [
-        "db.go",
+        "alias.go",
         "http_backup_handler.go",
-    ],
+    ] + select({
+        "//conditions:default": [
+            "db_kafka_wrapped.go",
+        ],
+        ":kafka_disabled": [
+            "db.go",
+        ],
+    }),
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/db",
     visibility = [
         "//beacon-chain:__subpackages__",
@@ -13,10 +27,14 @@ go_library(
     ],
     deps = [
         "//beacon-chain/db/iface:go_default_library",
-        "//beacon-chain/db/kafka:go_default_library",
         "//beacon-chain/db/kv:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
-    ],
+    ] + select({
+        "//conditions:default": [],
+        ":kafka_disabled": [
+            "//beacon-chain/db/kafka:go_default_library",
+        ],
+    }),
 )
 
 go_test(
